+++ app/includer/admins.py
@@ -1,5 +1,5 @@
 from entities.category.admins_category import CategoryAdmin
-from entities.service_name.admin_service_name import ServiceNameAdmin
+from entities.service_name.admins_service_name import ServiceNameAdmin
 from entities.users.admins_user import EmployeeAdmin, ClientAdmin
 
 admin_classes2 = (
